requirements.txt
asgiref==3.2.10
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-import-export
gunicorn==20.0.4
pytest==6.2.4
pytest-pythonpath==0.7.3
pytest-django==4.4.0
psycopg2-binary
PyJWT==2.1.0
pytz==2020.1
pytest-django==4.4.0
pytest-pythonpath==0.7.3
sqlparse==0.3.1
requests==2.26.0
python-dotenv
Pillow
djoser
pylint-django